+++ js/src/cli.ts
@@ -23,7 +23,7 @@ import path from 'path';
 import yargs from 'yargs'
 import {hideBin} from 'yargs/helpers'
 
-import {GoogleAdsApiClient, GoogleAdsApiConfig, loadAdsConfigFromFile} from './lib/ads-api-client';
+import {GoogleAdsApiClient, GoogleAdsApiConfig, loadAdsConfigFromFile, parseCustomerIds} from './lib/ads-api-client';
 import {AdsQueryExecutor, AdsQueryExecutorOptions, AdsApiVersion} from './lib/ads-query-executor';
 import {BigQueryInsertMethod, BigQueryWriter, BigQueryWriterOptions} from './lib/bq-writer';
 import {ConsoleWriter, ConsoleWriterOptions} from './lib/console-writer';
@@ -81,7 +81,7 @@ const argv = yargs(hideBin(process.argv))
   .option("account", {
     alias: ["customer", "customer-id", "customer_id"],
     type: "string",
-    description: "Google Ads account id (w/o dashes), a.k.a customer id",
+    description: "Google Ads account id (w/o dashes), a.k.a customer id or multiple accounts separeted with comma",
   })
   .option("customer-ids-query", {
     alias: ["customer_ids_query"],
@@ -312,16 +312,13 @@ function getWriter(): IResultWriter {
 }
 
 async function main() {
-  if (argv.account) {
-    argv.account = argv.account.toString();
-  }
   logger.verbose(JSON.stringify(argv, null, 2));
 
   let adsConfig: GoogleAdsApiConfig|undefined = undefined;
   let adConfigFilePath = <string>argv.adsConfig;
   if (adConfigFilePath) {
     // try to use ads config from extenral file (ads-config arg)
-    adsConfig = await loadAdsConfig(adConfigFilePath, argv.account);
+    adsConfig = await loadAdsConfig(adConfigFilePath);
   }
   // try to use ads config from explicit cli arguments
   if (argv.ads) {
@@ -331,15 +328,12 @@ async function main() {
       client_secret: ads_cfg.client_secret || '',
       developer_token: ads_cfg.developer_token || '',
       refresh_token: ads_cfg.refresh_token || '',
-      login_customer_id:
-          (ads_cfg.login_customer_id || argv.account || '')?.toString(),
-      customer_id:
-          (argv.account || ads_cfg.login_customer_id || '')?.toString(),
-    })
+      login_customer_id: ads_cfg.login_customer_id || '',
+    });
   } else if (!adConfigFilePath && fs.existsSync('google-ads.yaml')) {
     // load a default google-ads if it wasn't explicitly specified
     // TODO: support searching google-ads.yaml in user home folder (?)
-    adsConfig = await loadAdsConfig('google-ads.yaml', argv.account);
+    adsConfig = await loadAdsConfig('google-ads.yaml');
   }
   if (!adsConfig) {
     console.log(chalk.red(
@@ -348,9 +342,32 @@ async function main() {
   }
 
   logger.verbose('Using ads config:');
-  logger.verbose(JSON.stringify(adsConfig, null, 2));
+  logger.verbose(
+    JSON.stringify(
+      Object.assign({}, adsConfig, {
+        refresh_token: "<hidden>",
+        developer_token: "<hidden>",
+      }),
+      null,
+      2
+    )
+  );
 
-  let client = new GoogleAdsApiClient(adsConfig, argv.account);
+  let customerIds = parseCustomerIds(argv.account, adsConfig);
+
+  if (!customerIds || customerIds.length === 0) {
+    console.log(chalk.red(`No customer id/ids were provided. Exiting`));
+    process.exit(-1);
+  }
+  if (
+      !adsConfig.login_customer_id &&
+      customerIds &&
+      customerIds.length === 1
+    ) {
+      adsConfig.login_customer_id = customerIds[0];
+    }
+
+  let client = new GoogleAdsApiClient(adsConfig);
   let executor = new AdsQueryExecutor(client);
 
   // NOTE: a note regarding the 'files' argument
@@ -393,17 +410,18 @@ async function main() {
     logger.info(
       "Skipping account expansion because of disable_account_expansion flag"
      );
-    customers = [client.root_cid];
+    customers = customerIds;
   } else {
-    logger.info(`Fetching customer ids ${
+    // expand the provided accounts to leaf ones as they could be MMC accounts
+    logger.info(`Expanding customer ids ${
       customer_ids_query ? '(using custom query)' : ''}`);
-    customers = await client.getCustomerIds();
+    customers = await client.getCustomerIds(customerIds);
     logger.verbose(
-      `Customer ids from the root account ${client.root_cid} (${customers.length}):`
+      `Customer ids from the root account(s) ${customerIds.join(',')} (${customers.length}):`
     );
     logger.verbose(customers);
     if (customer_ids_query) {
-      logger.verbose(`Fetching customer ids with custom query`);
+      logger.verbose(`Filtering customer ids with custom query`);
       logger.debug(customer_ids_query);
       try {
         customers = await executor.getCustomerIds(customers, customer_ids_query);
@@ -452,9 +470,9 @@ async function main() {
       chalk.green('All done!') + ' ' + chalk.gray(`Elapsed: ${elapsed}`));
 }
 
-async function loadAdsConfig(configFilepath: string, customerId?: string|undefined) {
+async function loadAdsConfig(configFilepath: string) {
   try {
-    return loadAdsConfigFromFile(configFilepath, customerId);
+    return loadAdsConfigFromFile(configFilepath);
   } catch (e) {
     console.log(chalk.red(
         `Failed to load Ads API configuration from ${configFilepath}: ${e}`));