+++ js/src/lib/ads-api-client.ts
@@ -29,13 +29,13 @@ import { getLogger } from "./logger";
 export interface IGoogleAdsApiClient {
   executeQueryStream(
     query: string,
-    customerId?: string | undefined | null
+    customerId: string
   ): AsyncGenerator<any[]>;
   executeQuery(
     query: string,
-    customerId?: string | undefined | null
+    customerId: string
   ): Promise<any[]>;
-  getCustomerIds(): Promise<string[]>;
+  getCustomerIds(customerId: string): Promise<string[]>;
 }
 
 // export type GoogleAdsApiConfig = CustomerOptions&ClientOptions;
@@ -45,28 +45,22 @@ export type GoogleAdsApiConfig = {
   client_secret: string;
   developer_token: string;
   // CustomerOptions:
-  customer_id?: string;
   refresh_token: string;
   login_customer_id?: string;
   linked_customer_id?: string;
+  customer_id?: string[]|string;
 };
 
 export class GoogleAdsApiClient implements IGoogleAdsApiClient {
   client: GoogleAdsApi;
   customers: Record<string, Customer>;
   ads_cfg: GoogleAdsApiConfig;
-  root_cid: string;
   logger;
 
-  constructor(adsConfig: GoogleAdsApiConfig, customerId?: string | undefined) {
+  constructor(adsConfig: GoogleAdsApiConfig) {
     if (!adsConfig) {
       throw new Error("GoogleAdsApiConfig instance was not passed");
     }
-    customerId = customerId || adsConfig.customer_id;
-    if (!customerId) {
-      throw new Error(`No customer id was specified`);
-    }
-    customerId = customerId?.toString();
     this.ads_cfg = adsConfig;
     this.client = new GoogleAdsApi({
       client_id: adsConfig.client_id,
@@ -74,31 +68,22 @@ export class GoogleAdsApiClient implements IGoogleAdsApiClient {
       developer_token: adsConfig.developer_token,
     });
     this.customers = {};
-    this.customers[customerId] = this.client.Customer({
-      customer_id: customerId, // child
-      login_customer_id: adsConfig.login_customer_id, // MCC
-      refresh_token: adsConfig.refresh_token,
-    });
-    // also put the customer as the default one
-    this.customers[""] = this.customers[customerId];
-    this.root_cid = customerId;
     this.logger = getLogger();
   }
 
-  protected getCustomer(customerId: string | undefined | null): Customer {
+  protected getCustomer(customerId: string): Customer {
     let customer: Customer;
     if (!customerId) {
-      customer = this.customers[""];
-    } else {
-      customer = this.customers[customerId];
-      if (!customer) {
-        customer = this.client.Customer({
-          customer_id: customerId, // child
-          login_customer_id: this.ads_cfg.login_customer_id, // MCC
-          refresh_token: this.ads_cfg.refresh_token,
-        });
-        this.customers[customerId] = customer;
-      }
+      throw new Error("Customer id should be specified ");
+    }
+    customer = this.customers[customerId];
+    if (!customer) {
+      customer = this.client.Customer({
+        customer_id: customerId, // child
+        login_customer_id: this.ads_cfg.login_customer_id, // MCC
+        refresh_token: this.ads_cfg.refresh_token,
+      });
+      this.customers[customerId] = customer;
     }
     return customer;
   }
@@ -114,10 +99,7 @@ export class GoogleAdsApiClient implements IGoogleAdsApiClient {
       );
   }
 
-  async executeQuery(
-    query: string,
-    customerId?: string | null | undefined
-  ): Promise<any[]> {
+  async executeQuery(query: string, customerId: string): Promise<any[]> {
     const customer = this.getCustomer(customerId);
     try {
       return await customer.query(query);
@@ -127,10 +109,7 @@ export class GoogleAdsApiClient implements IGoogleAdsApiClient {
     }
   }
 
-  executeQueryStream(
-    query: string,
-    customerId?: string | undefined | null
-  ): AsyncGenerator<any[]> {
+  executeQueryStream(query: string, customerId: string): AsyncGenerator<any[]> {
     const customer = this.getCustomer(customerId);
     try {
       return customer.queryStream(query);
@@ -140,41 +119,72 @@ export class GoogleAdsApiClient implements IGoogleAdsApiClient {
     }
   }
 
-  async getCustomerIds(): Promise<string[]> {
-    const query =
-      `SELECT
+  async getCustomerIds(customerId: string|string[]): Promise<string[]> {
+    const query = `SELECT
           customer_client.id
         FROM customer_client
         WHERE
           customer_client.status = "ENABLED" AND
           customer_client.manager = False`;
-
-    let rows = await this.executeQuery(query);
-    let ids = rows.map((row) => row.customer_client.id!);
-    return ids;
+    if (typeof customerId === "string") {
+      customerId = [customerId];
+    }
+    let all_ids = []
+    for (const cid of customerId) {
+      let rows = await this.executeQuery(query, cid);
+      let ids = rows.map((row) => row.customer_client.id!);
+      all_ids.push(...ids);
+    }
+    return all_ids;
   }
 }
 
-export async function loadAdsConfigFromFile(
-  configFilepath: string,
-  customerId?: string | undefined
-): Promise<GoogleAdsApiConfig> {
+export function parseCustomerIds(customerId: string|undefined, adsConfig: GoogleAdsApiConfig) {
+  let customerIds: string[] | undefined;
+  if (!customerId) {
+    // CID/account wasn't provided explicitly, we'll use customer_id field from ads-config (it can be absent)
+    if (adsConfig.customer_id) {
+      if (_.isArray(adsConfig.customer_id)) {
+        customerIds = adsConfig.customer_id;
+      } else {
+        customerIds = [adsConfig.customer_id];
+      }
+    }
+  } else {
+    // NOTE: argv.account is CLI arg, it can only be a string
+    if (customerId.includes(",")) {
+      customerIds = customerId.split(",");
+    } else {
+      customerIds = [customerId];
+    }
+  }
+  if (!customerIds && adsConfig.login_customer_id) {
+    // last chance if no CID was provided is to use login_customer_id
+    customerIds = [adsConfig.login_customer_id];
+  }
+  
+  if (customerIds && customerIds.length) {
+    for (let i = 0; i < customerIds.length; i++) {
+      customerIds[i] = customerIds[i].toString().replaceAll('-', '');
+    }
+  }
+  return customerIds;
+}
+
+export async function loadAdsConfigFromFile(configFilepath: string): Promise<GoogleAdsApiConfig> {
   try {
     const content = await getFileContent(configFilepath);
     const doc = configFilepath.endsWith(".json")
       ? <any>JSON.parse(content)
       : <any>yaml.load(content);
+
     return {
       developer_token: doc["developer_token"],
       client_id: doc["client_id"],
       client_secret: doc["client_secret"],
       refresh_token: doc["refresh_token"],
-      login_customer_id: (doc["login_customer_id"] || customerId)?.toString(),
-      customer_id: (
-        customerId ||
-        doc["customer_id"] ||
-        doc["login_customer_id"]
-      )?.toString(),
+      login_customer_id: (doc["login_customer_id"])?.toString(),
+      customer_id: doc["customer_id"]?.toString(),
     };
   } catch (e) {
     throw new Error(